refactor: move private network hashing to helper file
